fix: archive resolve.c dead code, add WEB_ADMIN_TOKEN auth, fix -Wconversion warnings
- Move area/resolve.c to archive/resolve.c (legacy ident resolver, excluded from build)
- Add env-var token auth to webadmin/server.py: WEB_ADMIN_TOKEN env var, verify_token
  FastAPI dependency protects POST /api/wizinfo, /api/command, /api/backup, /api/shutdown,
  and WebSocket /ws/logs (token via query param)
- Fix all -Wconversion narrowing warnings across 17 C source files:
  act_comm, act_wiz, color, comm, db, edit, fight, handler, hunt, interp, magic, magic2,
  quest, save, special, update, wizlist
  Main patterns: (sh_int) casts on fread_number/affect field assignments, (int) casts on
  double-to-int float arithmetic, compound-assignment rewrites (xp += -> xp = (int)(xp*...))
- Build verified clean under both default flags and -Wall -Wextra -Wconversion

+++ src/comm.c
@@ -1501,7 +1501,7 @@ void write_to_buffer( DESCRIPTOR_DATA *d, const char *txt, int length )
      * Find length in case caller didn't.
      */
     if ( length <= 0 )
-        length = strlen(txt);
+        length = (int)(strlen(txt));
 
     /*
      * Initial \n\r if needed.
